requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
django-debug-toolbar==4.3.0
ipdb==0.13.13